Implement refresh timer
Implement refresh timer

News will be pulled from API only once per hour

+++ app/src/main/java/com/damir/stipancic/newsappv3/repository/NewsRepository.kt
@@ -7,25 +7,37 @@ import com.damir.stipancic.newsappv3.data.network.NewsApi
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 
+private const val ONE_HOUR_IN_MILLIS = 3_600_000
+
 class NewsRepository(private val database: ArticleDatabase) {
 
     suspend fun refreshNews(){
         withContext(Dispatchers.IO) {
-            val response = NewsApi.retrofitService.getNews()
-            response.body()?.let { newsResponse ->
-                newsResponse.articles.forEach{
-                    database.sleepDatabaseDao.insertArticle(it)
+            //check the age of the data in the DB and update if it's older than 1 hour
+            val latestEntry = database.sleepDatabaseDao.getLatestEntry()
+            if((System.currentTimeMillis() - latestEntry.createdAt) > ONE_HOUR_IN_MILLIS) {
+                val response = NewsApi.retrofitService.getNews()
+                response.body()?.let { newsResponse ->
+                    newsResponse.articles.forEach {
+                        database.sleepDatabaseDao.insertArticle(it)
+                    }
                 }
-            }
 
-            Log.d("newsRepository", "UNSAVED_COUNT: ${database.sleepDatabaseDao.getUnsavedRowCount()}")
-            if(database.sleepDatabaseDao.getUnsavedRowCount() >= 20)
-                database.sleepDatabaseDao.deleteNotNeededArticles()
-            Log.d("newsRepository", "UNSAVED_COUNT_AFTER_DELETE: ${database.sleepDatabaseDao.getUnsavedRowCount()}")
+                Log.d(
+                    "newsRepository",
+                    "UNSAVED_COUNT: ${database.sleepDatabaseDao.getUnsavedRowCount()}"
+                )
+                if (database.sleepDatabaseDao.getUnsavedRowCount() >= 20)
+                    database.sleepDatabaseDao.deleteNotNeededArticles()
+                Log.d(
+                    "newsRepository",
+                    "UNSAVED_COUNT_AFTER_DELETE: ${database.sleepDatabaseDao.getUnsavedRowCount()}"
+                )
+            }
         }
     }
 
     suspend fun getArticlesFromDB(): List<Article> {
-        return database.sleepDatabaseDao.getAllArticles()
+        return database.sleepDatabaseDao.getLatestArticles()
     }
 }
